refactor: add product package
