+++ src/main/java/kitchenpos/product/domain/JpaProductRepository.java
@@ -1,8 +1,7 @@
-package kitchenpos.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
+package kitchenpos.product.domain;
 
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
 }
