+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,11 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
 
+import java.math.BigDecimal;
+import java.util.UUID;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.math.BigDecimal;
-import java.util.UUID;
 
 @Table(name = "product")
 @Entity
