+++ src/test/java/kitchenpos/product/domain/InMemoryProductRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.application;
+package kitchenpos.product.domain;
 
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 
 import java.util.*;
 import java.util.stream.Collectors;
